Fix merge Co-authored-by: Antônio Lima <anttonio@usp.br> Co-authored-by: Artur Magalhães <artur_santos@usp.br> Co-authored-by: Diego de Morais <diegodemorais@gmail.com> Co-authored-by: Pedro Almeida <pedro.hba@usp.br> Co-authored-by: RenanSchwyz <renan.laiz@usp.br>
